Service endpoints now return appropriate HTTP status values

+++ src/main/java/com/airline/locationservice/controller/LocationControllerIata.java
@@ -1,11 +1,13 @@
 package com.airline.locationservice.controller;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import com.airline.locationservice.repository.AirportCodeIata;
 import com.airline.locationservice.repository.AirportCodeIataRepository;
-import com.airline.locationservice.repository.AirportCodeIcao;
+// import com.airline.locationservice.repository.AirportCodeIcao;
 
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,12 +17,15 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
 
 import com.delta.rm.core.location.AirportCode;
 import com.delta.rm.core.location.IATAAirportCode;
 
-// import com.delta.rm.core.location.AirportCode;
-// import com.delta.rm.core.location.IATAAirportCode;
+
+
+
 
 @RestController
 @RequestMapping("/airport/iata")
@@ -37,23 +42,31 @@ public class LocationControllerIata
     @GetMapping("")
     List<AirportCode> all()
     {
+        // Get list of airport codes from the DB, potentially an empty list.
         List<AirportCodeIata> result = repository.findAll();
+
+        if ( result.isEmpty() )
+        {
+            return new ArrayList<>();
+        }
+
+        // Map the results to Domain Objects rather than DB objects
         List<AirportCode> airports =
             result.stream().map( iata -> new IATAAirportCode(iata.getIataCode()) )
                            .collect( Collectors.toList() );
-
-        // airports.forEach( System.out::println );
 
         return airports;
     }
 
     @PostMapping("")
-    AirportCode newAirportCode( @RequestBody IATAAirportCode newAirportCode )
+    ResponseEntity<AirportCode> newAirportCode( @RequestBody IATAAirportCode newAirportCode )
     {
         AirportCodeIata iataDB = repository.save( new AirportCodeIata( newAirportCode.getAirportCode() ));
 
-        return new IATAAirportCode( iataDB.getIataCode() );
+        // Return 201 (Created)
+        return new ResponseEntity<>(new IATAAirportCode( iataDB.getIataCode() ), HttpStatus.CREATED );
     }
+
     // @PostMapping("/validate")
     // Boolean validateAirportCode( @RequestBody IATAAirportCode newAirportCode )
     // {
@@ -69,42 +82,46 @@ public class LocationControllerIata
     // Single item
 
     @GetMapping("/{id}")
-    AirportCode one(@PathVariable String id )
+    ResponseEntity<AirportCode> one(@PathVariable String id )
     {
         AirportCodeIata iata = repository.findById(id)
                                          .orElseThrow(() -> new AirportCodeNotFoundException( id ));
 
-        return new IATAAirportCode( iata.getIataCode() );
+        return new ResponseEntity<>( new IATAAirportCode( iata.getIataCode() ),
+                                     HttpStatus.OK );
     }
 
 
-    @PutMapping("/{id}")
-    AirportCode replaceAirportCode( @RequestBody AirportCodeIata newAirportCode, @PathVariable String id )
-    {
-        AirportCodeIata iata = repository.findById(id)
-            .map( airportCode -> {
-                airportCode.setIataCode(id);
 
-                return repository.save( airportCode );
-            })
-            .orElseGet(() -> {
-                newAirportCode.setIataCode( id );
-                return repository.save( newAirportCode );
-            });
+    @PutMapping("/{id}")
+    ResponseEntity<AirportCode> replaceAirportCode( @RequestBody AirportCodeIata newAirportCode, @PathVariable String id )
+    {
+        Optional<AirportCodeIata> iata = repository.findById(id);
+        if ( iata.isPresent() )
+        {
+            return new ResponseEntity<>( new IATAAirportCode( iata.get().getIataCode()), HttpStatus.OK );
+        }
 
         // Map the response to a Domain Object
-        return new IATAAirportCode( iata.getIataCode() );
+        return new ResponseEntity<>( new IATAAirportCode( repository.save( newAirportCode ).getIataCode()),
+                                     HttpStatus.CREATED );
     }
 
 
     @DeleteMapping("/{id}")
-    void deleteAirportCode( @PathVariable String id )
+    ResponseEntity<Boolean> deleteAirportCode( @PathVariable String id )
     {
-        repository.deleteById(id);
+        Optional<AirportCodeIata> iata = repository.findById(id);
+        if ( iata.isPresent() )
+        {
+            // HTTP 204 (No Content)
+            repository.deleteById(id);
+            return new ResponseEntity<>(Boolean.TRUE, HttpStatus.NO_CONTENT);
+        }
+        else
+        {
+            return new ResponseEntity<>( Boolean.FALSE, HttpStatus.NOT_FOUND);
+        }
     }
 
 }
-
-
-
-
